Adding pressure sensitivity.

+++ makeDeploy.py
@@ -1,14 +1,9 @@
 #!/usr/bin/env python
 import os
-import glob
 import shutil
-import zipfile
-
-def zipdir(path, ziph):
-    # ziph is zipfile handle
-    for root, dirs, files in os.walk(path):
-        for file in files:
-            ziph.write(os.path.join(root, file))
+import sys
+import getopt
+import platform
 
 def copytree(src, dst):
     for item in os.listdir(src):
@@ -27,18 +22,47 @@ def copytree(src, dst):
             shutil.copy(s, d)
 
 if __name__ == '__main__':
+    copyToBlenderAddons = False
+
+    for arg in sys.argv[1:]:
+        if arg == "-b":
+            copyToBlenderAddons = True
+
+    blenderHome = None
+    platSys = platform.system()
+    if platSys == 'Windows':
+        appData = os.getenv('APPDATA')
+        blenderHome = os.path.join(appData, "Blender Foundation/Blender/2.91")
+        
+    elif platSys == 'Linux':
+        home = os.getenv('HOME')
+        blenderHome = os.path.join(home, ".config/blender/2.91/")
+
+    #Create build directory
     curPath = os.getcwd()
     if os.path.exists('build'):
         shutil.rmtree('build')
     os.mkdir('build')
     os.mkdir('build/normalBrush')
-    
-    if os.path.exists('deploy'):
-        shutil.rmtree('deploy')
-    os.mkdir('deploy')
 
     copytree("source", "build/normalBrush");
 
-    shutil.make_archive("deploy/normalBrush", "zip", "build")
-
     
+    #Build addon zip file
+    if not copyToBlenderAddons: 
+        if os.path.exists('deploy'):
+            shutil.rmtree('deploy')
+        os.mkdir('deploy')
+
+        shutil.make_archive("deploy/normalBrush", "zip", "build")
+
+
+    if copyToBlenderAddons: 
+        addonPath = os.path.join(blenderHome, "scripts/addons")
+        destPath = os.path.join(addonPath, "normalBrush")
+
+        print("Copying to blender addons: " + addonPath)
+        if os.path.exists(destPath):
+            shutil.rmtree(destPath)
+        copytree("build", addonPath);
+
